+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - drop support of drupal 8.8 & 8.9
+- change bamboo_render_block to support block context-mapping
 
 ### Fixed
 - fix docker running tests on Github Actions